[ADAPTER] CID319159 Out-of-bounds write

+++ main/adapter/adapter.h
@@ -415,7 +415,7 @@ extern const uint32_t generic_btns_mask[32];
 extern struct bt_adapter bt_adapter;
 extern struct wired_adapter wired_adapter;
 
-uint8_t btn_id_to_axis(uint8_t btn_id);
+int32_t btn_id_to_axis(uint8_t btn_id);
 uint32_t axis_to_btn_mask(uint8_t axis);
 int8_t btn_sign(uint32_t polarity, uint8_t btn_id);
 void adapter_init_buffer(uint8_t wired_id);
